crypto/RPS/src/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pycryptodome==3.20.0